requirements.txt
speedtest-cli
pandas
openpyxl
schedule